requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface_hub
python-dotenv
PyPDF2


